# openapi.yaml
openapi: 3.0.3
info:
  title: Portfolio Dashboard API
  version: '0.1.0'
  description: Minimal API for the portfolio dashboard (LAN deployment)
servers:
  - url: http://localhost:3000/api
...
paths:
...
  /interest/months:
    get:
      tags: [Interest]
      summary: List interest months
      responses:
        '200':
          description: Interest months
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InterestMonth'
    post:
      tags: [Interest]
      summary: Add or replace interest month
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InterestMonth'
      responses:
        '201':
          description: Added interest month
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InterestMonth'
  /interest/months/{month}:
    delete:
      tags: [Interest]
      summary: Delete an interest month
      parameters:
        - name: month
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted
...
  /asset/{symbol}/history:
    get:
      tags: [Asset]
      summary: Get historical price data for an asset symbol
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
        - name: days
          in: query
          required: false
          schema:
            type: integer
            default: 60
      responses:
        '200':
          description: Asset history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetHistory'
  /history:
    get:
      tags: [History]
      summary: Get history points, optional range filter
      parameters:
        - name: range
          in: query
          required: false
          schema:
            type: string
            enum: [day, week, month, 6months, year, all]
            default: all
      responses:
        '200':
          description: History points
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/HistoryPoint'
    delete:
      tags: [History]
      summary: Clear all history points
      responses:
        '204':
          description: All history points cleared
  /history/point:
    post:
      tags: [History]
      summary: Add a history point
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HistoryPoint'
      responses:
        '201':
          description: History point created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HistoryPoint'
  /history/{id}:
    delete:
      tags: [History]
      summary: Delete a history point by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted
...
components:
  schemas:
...
    HistoryPoint:
      type: object
      properties:
        id:
          type: string
        t:
          type: string
        ts:
          type: integer
        v:
          type: number
        i:
          type: number
        p:
          type: number
        manual:
          type: boolean
        note:
          type: string
...
    InterestMonth:
      type: object
      properties:
        month:
          type: string
        amount:
          type: number
...
    AssetHistory:
      type: object
      properties:
        symbol:
          type: string
        labels:
          type: array
          items:
            type: string
        prices:
          type: array
          items:
            type: number
        ts:
          type: integer
...